Move auth theme files under feature scope. Isolates styling per feature.
